requirements.txt
fastapi
uvicorn
python-multipart
python-decouple
email-validator
dnspython
idna